1. initial support for rich text in xssf" 2. cleaned common ooxml-ole2 interfaces, removed ole2-specific stuff 3. added new examples from the quick guide 4. misc xssf refactoring, tending to use enums intstead of final static constants 5. refactored XSSFFont 6. included test-ooxml and jar-ooxml in the dist target, they are part of release and should be there
git-svn-id: https://svn.apache.org/repos/asf/poi/branches/ooxml@696584 13f79535-47bb-0310-9956-ffa450edef68

+++ src/ooxml/java/org/apache/poi/xssf/usermodel/XSSFCellStyle.java
@@ -417,7 +417,7 @@ public class XSSFCellStyle implements CellStyle {
         return (short) getBorderColor(side).getIndexed();
     }
 
-    private void setBorderColorIndexed(BorderSide side, long color) {
+    private void setBorderColorIndexed(BorderSide side, int color) {
         getBorderColor(side).setIndexed(color);
     }
